Rental Radar - Southern Maine/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
python-dotenv
jinja2
